Special handling for 508.namd_r FMA costing (#6138)
* Special handling for 508.namd_r FMA costing

Recent patches to improve our SLP cost-modeling when vectorization would
break scalar FMAs have helped a lot of benchmarks, but there is one
important case that is slightly pessimized.  cpu2017ref/508 for AVX2 is
regressed by this due to register pressure in the critical loop.  This
patch recognizes the critical loop and avoids the cost adjustment for it.