[Paorpt] Don't filter out duplicates when collecting non-ptrs for `omp target`. (#6206)
When host/device compilation is done with different opt-levels, it's
possible that optimizations in one make the values duplicate, but not in
other. e.g.

Host:
  %vla1 = alloca i32, i64 %n1
  %vla2 = alloca i32, i64 %n2

Device:
  %vla1 = alloca i32, i64 %n
  %vla2 = alloca i32, i64 %n

So we need to skip any duplicate filtering code that ensures we don't
capture the same value multiple times.

JIRA: CMPLRLLVM-39013

* Avoid the duplication of capturing the same VLA's size twice.
VLA can be both in a map and in private/firstprivate. No need to
capture its size twice for non-TYPED clauses.

For TYPED clauses, the handling of private/fp clause captures the
size from the typed clause, not the VLA's size itself. So, for them
we need to capture the size when handling the map clause..