cmplrllvm-38879: emit firstprivate for variable with is_device_ptr in (#6217)
implicit task.

This is requset from BE.

For variable with map, it is required to shared in implicit task.

But not for is_device_ptr. Since we now generate map for is_device_ptr
The change is just to skip generate shared in implicit task for variable
with is_device_ptr.