CMPLRLLVM-37557: Add a threshold for the number of If instructions that could be unswitched (#6101)
* CMPLRLLVM-37557: Add a threshold for the number of If instructions that could be unswitched

This patch adds a threshold for regions that were modified and
contains a large number of IF instructions. If the loop that is
being unswitch contains a large count of If instructions then
we will stop the unswitch. The reason is that it increases the
compile time while it won't be benefitial to the generated code.

* Add comment addressing the else part.

* Fix for CJ's comment

* Fix for clang format

* Relax conditions to only skip when if and else are in the loop nest.