CMPLRLLVM-39052: Inhibit limit on inlining of single callsite linkonce ODR functions under SYCL. (#6277)
* CMPLRLLVM-39052: Inhibit limit on inlining of single callsite linkonce ODR
functions under SYCL.

This change will revert the change to the inlining heuristic for limiting
the inlining of single callsite linkonce ODR functions for SYCL programs.
This is accomplished using the -mllvm -sycl-opt for device compilations and a
new backend option -mllvm -sycl-host for host compilations. Mike Toguchi
provided the changes to the driver. I talked to Andy Kaylor about where to put
the cl::opt for sycl-host and he said it was all right to put it in the
Inliner files because it is currently only needed by the inliner.

* LIT test fix from Mike Toguchi