[VPlan][inscan] Keep inscan reduction initialization unmasked in masked loop (#6272)
Inscan reduction implementation relies on having the reduction
reinitialized in the loop header:

pre.header:
  float* %red = allocate-priv float*
  float %red_load = load float* %red
  float %red_init = reduction-init-scalar float 0.0 float %red_load
header:
  float %carry-over =
      phi float [ %red_init, pre.header ]
                [ %extract_last_lane, separating.pragma.block ]
  store float -0.000000e+00 float* %red
separating.pragma.block:
  float %red_load_run = load float* %red
  float %running_red = running-inclusive-reduction float %red_load_run
                                                   float %carry-over
                                                   float -0.000000e+00
  store float %running_red float* %red
  float %extract_last_lane = extract-last-vector-lane float %running_red
post.exit:
  float %red_pre_final = load float* %red
  float %final_red = reduction-final-inscan float %red_pre_final
For masked mode loop the condition is created before the first non-phi
instruction in the header block meaning that the reinitializing store
will also be masked. This situation would lead to garbage in the masked
out lanes, so the store needs to be hoisted out of the masked mode loop
condition:

header:
  float %carry-over =
      phi float [ %red_init, pre.header ]
                [ %extract_last_lane, separating.pragma.block ]
  store float -0.000000e+00 float* %red
  br %masked_kernel
masked_kernel:
...

Amend VPAllocatePrivate with a field indicating Entity code that this
alloca was created for. This way the initialization store for inscan
reduction will be the first one pointer operand being alloca with
inscan reduction opcode.

Keeping Entity opcode in VPAllocatePrivate would also help to overcome
CMPLRLLVM-38919, where loads from in-memory unit-step induction
lead to gather/scatter as the step information is lost.