(xmain-rel) [AndersensAA] (CMPLRLLVM-38623) Propagate UniversalSet to points-to sets of all dereference levels of a pointer If UniversalSet is in point-to set of the pointer. (#6178)
* (xmain-rel) [AndersensAA] (CMPLRLLVM-38623) Propagate UniversalSet to points-to sets of all dereference levels of a pointer If UniversalSet is in point-to set of the pointer.

While solving constraints, if UniversalSet is in points-to set of CurrNode,
it propagates UniversalSet to the other deference levels of the pointer (if
it is not already there).

Fixed some LIT tests that fail due to this change.

This causes performance regression:
EEMBC regression: Need to handle user routine t_printf as wrapper for vsprintf so that arguments of t_printf calls are not treated as conservative. Zperf shows some gains and losses for EEMBC but they are not real.

spec_omp2012ref/371: Handled AbstractCalls to improve points-to but still there
is perf regression (8%) in zperf. Analyzed this regression. It doesn't look like it is real.

* Fix performance regression (CMPLRLLVM-39043). Combine https://github.com/intel-restricted/applications.compilers.llvm-project/pull/6235