[SLP] Do not reorder operands before adding to MultiNode. (#6376)
This patch is primarily intended to save a bit of compile time
and also simplifies a bit handling of MultiNode operands as we
don't need to seek for operand number for frontier instructions.
Since we reorder MultiNode operands anyway, we don't really have to
do extra reordering prior to adding operands to a MultiNode. This is
not a strictly NFC change as it may affect horizontal ordering (i.e.
which scalar element goes into which lane). Besides that the patch
is not supposed to change SLP vectorizer behavior.