Don't clone loop defining tokens inside which has user outside (#6364)
* Don't clone loop defining tokens inside which has user outside

CMPLRLLVM-38833: Tokens can't be used in PHI so LCSSA won't create
PHI in exit BBs for them. Therefore we can't clone this kind loop in
normal way. Hetero arch opt will try to clone normal ancestor loop
that near them or refuse to optimize such candidate loop if it and its
parent's are all not cloneable.

* Add lit test

* Clear NoCloneLoops for each function

* Fix Path order

* Also clone exit BB if it's "terminator" BB that use token defined in loop

* Rename lit test

* Don't need to fix phi for cloned exit BB

* Fix bug