[Driver] Introduce Windows aliases for -m<code> options (#6390)
Traditionally, we have supported the -m<code> feature enabling options on
Linux/macOS for the Classic compiler. Moving to the LLVM based compiler,
the driver has enabled the -m<code> options on Windows via the 'core option'
extensions in the driver. It would make sense to sync up Linux and Windows
here to provide the -m<code> ability on Windows.  This is covered by the
'/Qm<code>' syntax

We generally cover all x86 target options, with a few additional general
options which are used for various hardening needs.