[InstCombine] Remove stack save/restores in additional cases (#6502)
* [InstCombine] Remove stack save/restores in additional cases

Inlining into an OpenMP region can result in unused stacksave/stackrestore pairs which are desirable to remove. Often there is a lifetime for the allocation as well.

We already have a customization intended to allow this removal by checking that any intermediate intrinsic call has no side effects.  However, @llvm.lifetime intrinsics are considered to potentially store to their arguments, so this is insufficient.  The accompanying test actually tests the case with an intermediate lifetime but passes as a special case handled directly by upstream when the following terminator is a 'ret'.

This change loosens the customization to allow intermediate intrinsics as long as they are IntrinsicInst::isAssumeLikeIntrinsic, and updates the custom LIT test to exercise the customization as originally intended.

Toward Jira CMPLRLLVM-36536.