[NFC] CMPLRLLVM-38435: Framework for multiple Structs per memfunc (#6515)
* [NFC] CMPLRLLVM-38435: Framework for multiple Structs per memfunc

For 525.x264, there is a key structure where one field is a nested structure,
and only part of it is written.  The current xmain implementation goes
conservative in this case. This is a first step to extend the infrastructure
to handle an instruction that could set parts of multiple Structs. This is
a non-functional change at this point, as the recognition is not yet
implemented or enabled.

* Fixes for Chris's commments

* Fix for Chris's comment