[NFC][LLD][COFF] Temporary fix to prevent that exported symbols aren't modified (#6517)
* [NFC][LLD][COFF] Temporary fix to prevent that exported symbols aren't modified

The structure Config::Export use StringRefs to handle the fields used for
storing symbol name, etc. In some cases the fields are written using a
StringSaver object, which basically copies the actual string pointed by a
StringRef to an independent memory space and return the new StringRef. Other
cases it is just change the address where the pointer is pointing to. This
patch makes sure that StringRef fields in Export have their own memory
space allocated rather than sharing with other structures.

* Fix for copyright