[VecClone][VPlan] Fix generation of variants with vector-of-bool (#6519)
parameters.

Previously, if a function was annotated with `#pragma omp declare simd`
and also already had parameters that were of vector-of-bool (`<N x i1>`)
type, e.g:

  typedef bool bool2 __attribute__((ext_vector_type(2));

  #pragma omp declare simd simdlen(16)
  void foo(bool2) {}

We would incorrectly generate the corresponding vector variants: without
promoting `<N x i1>` to `<N x i8>`; which led to a mismatch when
generating calls to the vector variant: as though the arguments *were*
promoted. This led to a crash during VPlan CG.

The fix is to simply promote `<N x i1>` parameters to `<N x i8>` in the
same manner as we do with `i1` params to `i8`. This results in the
correct variant generation and fixes the crash in CG.