[Driver] Move default configuration file processing into clang driver (#6516)
When using 'icx' or 'icpx' the configuration file processing:
  - discovering the proper .cfg name
  - finding the file in the default bin location
  - using environment variables to set default cfg used
was being done in the 'icx' or 'icpx' wrapper before the information
was passed to the clang driver via the --intel-config option.  Effectively
move these behaviors into the clang driver, reducing the responsibilities
of the wrappers.  This also introduces DPCPPCFG and the corresponding
dpcpp.cfg default configuration file that can be used with the dpcpp
driver.