[VPlan][HIR] Add code generation support for VectorVariants based call vectorization in VPlan HIR vectorizer. (#6385)
* [VPlan][CodeGenHIR] Fix for CMPLRLLVM-37334

This patch adds support to vplan to vectorize loops containing calls to
externed simd functions and also to call the vectorized function instead
of scalarizing the calls. These fixes are for HIR as LLVM side of things
is already done. Currently, pumping support and use of masked variant
using all-ones mask for unmasked is not available.

* [VPlan][HIRCodeGen] Fix for CMPLRLLVM-37334

Add test for vector function defined in same module

* [VPlan][CodeGenHIR] Fix test

Since vectorizing loops with vector functions is now supported for HIR,
this test started failing because the calls to the vector functions were
scalarized due to VF=2 and no variant was available for that VF. Also,
things were further complicated by the fact that the loop was completely
unrolled after vectorization and only the scalarized calls remained. The
test originally looks at IR dumped just after VPlanDriverHIR, and this is
before the LLVM is cleaned up so the calls were appearing in the entry
block with the loop still present in the dump. The HIR region was shown
correctly (e.g., no loop). It isn't until later when hir-cg creates
valid LLVM-IR. Instead of adding more options and changing the dumps, it
was easier to just keep the original intent of the test by forcing VF=1
and thus preventing the unrolling and scalarization of the vector function
calls.

* One more test fix

* Follow-up fixes:
1) Generate the correct mask type for masked vector variants
2) Support using masked variant with all-ones mask when unmasked
   variant is not available

* Add comment to test

* Remove old comment and update test

* Fix prod build issue for variable used only in assert