[SLP][NFCI] Sink DisableReorder flag down into MNOperands class. (#6573)
The problem concerns path steering for the case when MultiNode operands
are not legal to reorder. The main discrepancy here is that we call
reorder() method to collect operands scores but the method does actually
perform reordering and collects scores for operands as if they were reordered.
If the flag was set, we discarded that reordering and only used operands scores
to steer vectorization. When reorder() function did not happen to reorder the
operands than it had perfect sense to use the scores for steering. Otherwise
we were applying scores of operands, collected as if the operands were reordered,
onto their original order. This patch addresses specifically such cases.
For perfectly vectorizable operands that do not need any reordering this
patch is NFC. For cases where scores were collected incorrectly the patch may
change order of visiting operands but the outcome should be nearly the same.