CMPLRLLVM-39072: Disable codegen in unswitch for non-innermost loop if it won't benefit another transformation (#6588)
* CMPLRLLVM-39072: Disable codegen in unswitch for non-innermost loop if it won't benefit another transformation

This change set will collect the if cases that will be unswitched due to
relaxing the condition for non-innermost loops. If the conditions are the
only candidates for unswitching, then the loop won't be set for code gen.
If unswitching a non-innermost loop benefits another transformation
(e.g. unroll and jam), then this transformation will enable the code gen
and the transformation will be applied.

* Fix for Pankaj' comments. Also remove the incorrect test case.

* Add const to the function