[OpenMP] Parsing/Sema/Serialization of 'interop' clause (#6616)
* [OpenMP] Parsing/Sema/Serialization of 'interop' clause

This is a proposed clause for OpenMP 6.0. Since 6.0 isn't done
it isn't conditioned on that yet (that macro value must be known
before we do that). Limit it to 5.1 since 'omp dispatch' was introduced
there.

The clause uses the same type of interop variable as the interop directive
init/use/destroy clauses and uses the same checking.

Partially implements CMPLRLLVM-37338.