CMPLRLLVM-38435: Eliminate 'bad casting' when a pointer from a struct is used in a scalar access, but retain 'mimatched element access'. (#6718)
* CMPLRLLVM-38435: Eliminate 'bad casting' when a pointer from a struct
is used in a scalar access, but retain 'mimatched element access'.

There is a place in 525.x264 where a i8* pointer from struct x264_t is
accessed and passed down to a call site as a i16* pointer. In these cases,
there is no accessing of a struct using the pointer. Chris agreed that this
could be treated as a 'mismatched element access' rather than as a 'bad
casting'. This is significant because 'bad casting' is a cascading safety
check, which 'mismatched element access' is not. We want to avoid cascading
down to 'struct mc_functions_t' which includes the key function pointers
in this benchmark.

* Fixes for Chris's comment

* Fix for LIT test condition noticed by Chris

* clang formatting of code requested by Chris

* Fix for Clang Format, noted by Chris

* Clang format