[VPO] Introduce mechanism to prohibit memory motion of OMP SIMD clause variables. (#6518)
While testing support for UDRs in xmain vectorizers, it was
observed that in many cases the pointer value that represents
given UDR is hoisted/moved out of the loop which eventually
leads to registerization of this variable. This is not
conducive for downstream middle-end passes as additional alias
analysis is needed to correctly identify usage of clause
variables in target loop. This patch introduces a new mechanism
in Paropt sub-component to prohibit cross-loop motion of these
pointers. This is achieved by introducing a new memory motion
guard work region (WRNGuardMemMotion) and capturing the pointers
of interest as live-in values. The guards are appropriately
inserted in target loop's header and latch BBs. Renaming of these
captured pointers is also done to prevent motion of any GEPs
operating on them. This is the first patch in the series for
the task to avoid registerization of UDRs in xmain middle-end.

Change-Id: I4ddcc456e7f9f63bd91d6f34c68ce8cf957029fa