CMPLRLLVM-39668: Implement #pragma forceinline recursive and #pragma (#6784)
inlinehint recursive for the new pass manager.

I left the old pass manager test lines into the LIT tests for now since we
never had LIT tests for the backend implementation. These pragmas were
implemented before we had the requirement to put in LIT tests, so we did
not detect the loss of the implementation when we moved to the new pass manager
using LIT tests.