[NFC] Remove -vplan-enable-new-cfg-merge[-hir] flags from lit tests (#6851)
* [NFC] Remove -vplan-enable-new-cfg-merge[-hir] flags from lit tests

The "new CFG merge" support has been stable and enabled by default for
some time now, and it's time to remove the temporary flags that allow
it to be turned on and off.  In preparation for that, this patch removes
the flags from lit tests.  In many cases this leads to duplicate RUN
lines that have been consolidated.  This is a large but straightforward
and mechanical patch.

I'd like to call the reviewers' attention to two tests:

  Transforms/Intel_VPO/Vectop/hir_zext_wrap.ll
  Transforms/Intel_VPO/Vecopt/hir_vplan_loop_unroll_pragma.ll

For hir_zext_wrap.ll, the IGNORE-WRAP lines were previously only checked
with the new CFG merge support disabled.  When I removed the flag (thus
enabling the support), the HIR changed slightly.  I don't know enough
about this test to know whether that change is expected, so please check
that.

Similarly, all of the RUN lines for hir_vplan_loop_unroll_pragma.ll
disabled the new CFG merge support.  In this case removing the flag
caused really substantial changes to the HIR.  It's not obvious to me
what the proper modification to the test should be.  For the time
being I've removed the check-prefix for VPVALCG from the RUN lines
so that the test passes, but I'd like advice about what to do with
this particular test.  (FWIW, I tried running intel_update_vplan_checks.py
on it, but the utility took a runtime exception.)

The rest of the changes in this patch are straightforward.

* [NFC] Additional changes from code review

I updated comments in several tests in the HIROptReport subdirectory,
and modified the two tests I asked about per Satish's suggestions.