[Driver] Update Windows icx behavior to allow for Windows options (#6850)
Update icx on Windows to allow for Windows options as well as the standard
Linux based options.  When a user specifies any MSVC specific options when
using icx emit a diagnostic that the user should be using icx-cl instead.
This is a transitional change as we introduce the gamut of icx/icpx/icx-cl to all
of the OSs.