[OpenCL] Fix kernel stack usage estimation (#6883)
The kernel metadata 'barrier_buffer_size' records the required barrier
buffer size for each work item. When kernel is vectorized or
subgroup-emulated, we may store a <VF x Ty> vector to the barrier buffer
in the last iteration of barrier loop, even if only part of work items
are active (WG_Size % SG_Size != 0).

So the actual barrier buffer size allocation (PrepareKernelArgs pass)
would be a round-up to the multiple of subgroup size. This may waste
memory in some cases.

For example, the kernel is vectorized to VF=16, but the local sizes are
{1,1,1}. Say we need 1MB barrier buffer for each work item, then the
actual barrier buffer allocation would be 16MB because the local size on
vectorization dimension is rounded-up to 16 (multiple of VF).

Previously we didn't take this rounding-up into consideration of the
kernel stack size, so we were under-estimating kernel stack usage. As a
result, auto-memory sometimes doesn't take effect when it should, the
application crashes because of stack overflow.

This patch fixes the calculation of kernel stack size (especially the
barrier buffer size) so that auto-memory can opt-in to always provide a
sufficient stack frame for kernel execution.

We may investigate further to see if we can avoid wasting barrier buffer
memory for these corner cases.

CMPLRLLVM-39723