[NFCI] Remove -vplan-enable-new-cfg-merge[-hir] flags and clean up dead source (#6876)
* Remove -vplan-enable-new-cfg-merg[-hir] flags and clean up dead source

The "new CFG merge" support has been stable and enabled by default for
some time now, and it's time to remove the temporary flags that allow
it to be turned on and off.  A previous patch cleaned up lit tests so
the flags are no longer in use.  This patch removes the flags themselves,
and simplifies code that was dependent on their setting.

As a result of this, the VPOCodeGenHIR::isMergedCFG() interface was
reduced to a check for !isSearchLoop().  To clarify this and expose
any possible misuses, I've replaced all uses of the former with the
latter.  Presumably at some point we can do less special handling
for search loops and further simplify the code.