[VPlan][inscan] Add support for partially registerized inscan reduction (#6852)
When registerized, inscan reduction becomes partially registerized,
it has the recurrence phi used in the intput phase, and memory is used
in the scan phase (scan directives prevent full registerization).
The load from the reduction memory would be live out, not the phi.
For registerized inscan reduction the recurrence phi needs to be
replaced with identity value.

The replaced phi could be removed from VPlan as it's dead,
however, further assumptions imply having recurrence phi for liveouts,
unless it is private unconditional. Since the phi is dead, it makes more
sense to leave it in the VPlan, rather than special case the assumptions.