[Paropt][Offload] Support INTEROP clause for DISPATCH construct (#6899)
JIRA: CMPLRLLVM-37340

When INTEROP clause is present, pass its argument as the interop obj
to the variant function.
If the DEVICE clause if absent, obtain the device num from the interop obj.
Current implementation supports one interop obj in the INTEROP clause.
This is sufficient for now. We will support list of interop objs in the future.