[CMPLRLLVM-39584] Fix crash during global DCE (#6923)
* [CMPLRLLVM-39584] Fix crash during global DCE

An IR change made by the attributor pass was causing a compiler failure to
occur during the dead code elimination pass. The problem was triggered
because the Attributor performed constant folding on a GEP with an address
space cast to determine a replacement value.

When the attributor was performing replacements, it tried to apply the
following change:

Use       : double addrspace(3)* getelementptr inbounds
               (%struct, %struct addrspace(3)* @counter, i32 0, i32 0, i32 0)

in        : double addrspace(4)* addrspacecast
              (double addrspace(3)* getelementptr inbounds
	        (%struct, %struct addrspace(3)* @counter, i64 0, i32 0, i32 0)
	      to double addrspace(4)*)

instead of: double addrspace(3)* getelementptr inbounds
              (%struct, %struct addrspace(3)* @counter, i64 0, i32 0, i32 0)

This subepxression that is being replaced is a ConstantExpr that is
within another ConstantExpr. All ConstantExpr values are stored in
a hash table and can be shared among multiple instructions to use the same
ConstantExpr across the module. Ddirectly replacing part of the expression will
corrupt the hash table becuase the new value will hash to a differnt value than
the original, as well as potentially change functions beyond the one the Attributor
is currently working on.

When a lookup is performed for the new ConstantExpr, such as from global DCE when
the last instance is removed from the IR, a failure would occur.

This change inhibits the Attributor from performing this replacement of part
of a ConstantExpr.

[CMPLRLLVM-39584]

* Update copyright