[Driver] Fix issues with default pie settings for device compilations (#6739)
When performing device compilations with -fsycl, the device objects
are wrapped.  These wrapped objects should be built in the same manner
as the host object.  We use 'llc' to do this step, which was neglected
in the pursuit for default PIE behaviors.  Do some refactoring

Adjust behaviors to only link in the crt* objects in accordance to PIE settings.
This is different than the original change, which changed the compilation
model to be -fPIE -pie across the board.