[CMPLRLLVM-39884][DTrans] Handle Expand Parameters for i386 (#6943)
* [CMPLRLLVM-39884][DTrans] Handle Expand Parameters for i386

When developing the DTrans metadata information handling, a reproducer
for Expand parameters wasn't obvious, so it was left as an assert.

CMPLRLLVM-39884 reported an example of how to reproduce this. Expand
simply 'flattens' a struct into parameters if it meets some
requirements. This patch ensures that the metadata is properly generated
for these cases.