[NFC] Changing remark #15553 (#6937)
* [NFC] Changing remark #15553

We currently do not support auto-vectorization of outer loops.
Originally this remark gives misleading vect message:
"at O3 outer loop is not an auto-vectorization candidate at -O2. Consider using -O3."
But O3 was used. So this commit changes remark to
"loop was not vectorized: outer loop is not an auto-vectorization candidate." for
more proper message given

JIRA: CMPLRLLVM-33618

* Changed lid test.