Read machine loop location from loop metadata (#6924)
This is a small change to improve the accuracy of machine loop debug
locations by leveraging our ability to track down IR loop metadata and
reading the debug location from there if present. This should have
minimal impacts on codegen, but it does significantly improve the
accuracy of debug locations printed by the MIR opt-report emitter.