CMPLRLLVM-38435: Normalization of PHI inputs for 525.x264 (#6951)
* CMPLRLLVM-38435: Normalization of PHI inputs for 525.x264

The inputs of PHIs need to be normalized for 525.x264 so that the inputs have the same ptr type.
Without doing this, the element pointees of the ValueTypeInfos will not be consistent. Sometimes
they will include structure elements, sometimes not. This can lead to safety check violations.

For example, consider:
   %struct.Y = type { [7 x ptr] }
   %struct.X = type { i64, %struct.Y }
   %x1 = getelementptr inbounds %struct.X, ptr %arg, i64 0, i32 1
   %y2 = getelementptr inbounds [7 x ptr], ptr %w, i64 0, i64 0
   %63 = phi ptr [ %x1, %B0 ], [ %y2, %B1 ]
after normalization this will be:
   %x1 = getelementptr inbounds %struct.X, ptr %arg, i64 0, i32 1
   %dtnorm = getelementptr inbounds [7 x ptr], ptr %x1, i64 0, i32 0
   %y2 = getelementptr inbounds [7 x ptr], ptr %w, i64 0, i64 0
   %63 = phi ptr [ %dtnorm, %B0 ], [ %y2, %B1 ]
so that the source element type of the two PHI inputs is consistent.

* Fix comment