Revert "CMPLRLLVM-38691: Convert select instructions into if/else to enable loop unswitching (#6628)" (#6974)
Reverts the following commit due to compile time regression detected during post commit testing on cpu2017ref_507 on Windows.

PR: #6628 -

CMPLRLLVM-38691: Convert select instructions into if/else to enable loop unswitching

This change set enables converting select instructions into if/else if the
condition is loop invariant and the select instruction is the in the
innermost loop. Once the select instruction is converted into if/else,
the new conditional will be added as a candidate for unswitching, which
will do the final multiversioning.