Add flag to llvm-config to report opaque pointer mode that xmain was built as (#6961)
To support the transition to opaque pointer mode in xmain, for testing
we need a way to mark LIT tests that are not expected to work under
opaque pointer mode, and will be deleted once everything is switched
to opaque pointers. For example, in DTrans there is a set of tests
that were using the legacy DTrans infrastructure that are not expected
to work if the compiler solely uses opaque pointers. A set of new tests
that replace them has been developed to test the replacement DTrans
infrastructure.

To support testing of opaque pointer, without a lot of expected LIT
test fails, we need a way of marking tests in the backend as:
"UNSUPPORTED: enable-opaque-pointers", similar to way Clang tests
have this ability.

This PR introduces the option to llvm-config that reports the value
used during the CMake step for the ENABLE_OPAQUE_POINTERS flag. One
DTrans test as an example of the its use is included here. After this
change is merged, additional changesets will follow that mark other
LIT tests for the legacy DTrans infrastructure.

[CMPLRLLVM-39919]