Prevent PtrTypeAnalyzer assertion from being hit when analyzing subtract of two constants (#6973)
The PtrTypeAnalyzer was hitting an assertion because it was trying to
get the ValueTypeInfo object for an operand of a subtract instruction
that is a compiler constant using a method that requires a Value object
which is not a constant. It was not epxected that there would be a
subtract of the form that had 2 constnat operands. This optimizes the
visiting of a subtract instruction to avoid entering the analyzeValue
path when the subtract is just 2 constants, and makes sure the incorrect
method is not used in the case that the subtract is analyzed due to
it being a dependency for another instruction.

[CMPLRLLVM-39915]