[CMPLRLLVM-39914][DTRANS] Fix DTrans info for incomplete fptr & bitfields (#6969)
As reported in CMPLRLLVM-39914, we weren't properly generating
dtrans information in 2 cases:
1- A function pointer with an incomplete parameter, which just results
in an empty literal struct.
2- Bitfields that are combined into an array.

The former as is easy as just not adding fields when we detect this
situation, and was protected with an assert.  The latter is fixed by
detecting this situation and 'fixingup' the array correctly to make sure
we generate correct metadata.