CMPLRLLVM-38691: Convert select instructions into if/else to enable loop unswitching (#6982)
This change set enables converting select instructions into if/else if the
condition is loop invariant and the select instruction is the in the
innermost loop. Once the select instruction is converted into if/else,
the new conditional will be added as a candidate for unswitching, which
will do the final multiversioning.

This patch relands PR #6628, with a fix for the compile time in
CPU2017/507.cactus on Windows.