Fix crash from opaque pointer struct GEP in HIRRowWiseMV (#6872)
HIRRowWiseMV queries DTrans analysis to find values that may be assigned
to every element in a row of an array. This can only be done when the
array's base pointer is a struct field and HIRRowWiseMV needs to
determine which struct field it is to do the query. In the past with
typed pointers, if it found a struct GEP for that base pointer it could
rely on the third operand being the index into the struct:

    %a_ptr = getelementptr inbounds %struct.A_t, %struct.A_t* @A, i32 0, i32 0
    %a = load double*, double** %a_ptr

However, with opaque pointers the GEP does not need that third operand
if the field is the first one (field zero) because its pointer is the
same as the pointer for the whole struct:

    %a_ptr = getelementptr inbounds %struct.A_t, ptr @A, i32 0
    %a = load ptr, ptr %a_ptr

This caused a crash as HIRRowWiseMV attempted to access the third
operand. Instead of trying to access the nonexistent third operand in
this case, it should just do the query with field zero instead.

Jira: CMPLRLLVM-39742