CMPLRLLVM-39892: Stop assert due to trying to remove LibFunc (#6983)
* CMPLRLLVM-39892: Stop assert due to trying to remove LibFunc

In xmain, a LibFunc can have IR. Such a LibFunc could be converted into a
LibFunc without IR by the Attributor. The lazy call graph requires that
libFuncs not be removed as dead functions. So, do not put them on either of
the dead function lists in these cases.

* The pulldown restricted the definition of the library function printf to
always return an int, even though some versions return void. The library
functions will need to be generalized, but for now, we can change the
definition of printf to return an int in these new LIT tests so that
they illustrate the issue that needs to be fixed in this JIRA item.