CMPLRLLVM-38435: Multi-struct memfunc handling for 525.x264. (#7068)
* CMPLRLLVM-38435: Multi-struct memfunc handling for 525.x264.

This extends memfunc handling to get info on reads and writes that require
multi memcallinfos to describe the reading and writing of a structure with
nested structures. This is done to support field single/multiple value
analysis. In particular, it allows us to conclude that in the mc_functions_t
structure in 525.x264, the @mc_chroma, @mc_luma, and @get_ref functions are
the only non-nullptr values in 3 fields of that structure.

A new safety check 'Multi-struct memfunc' is introduced to inhibit
optimizations like field deletion and field reordering which may need to
recode memfuncs.

* Fixes requested by Chris