[VPlan][HIR] Generate VPSubscriptInst for memory/addressof refs always during HIR decomposition (#7048)
* [VPlan][HIR] Generate VPSubscriptInst for memory/addressof refs always during HIR decomposition

With recent work to make sincos transformation explicit in VPlan HIR path, we
need to assign proper symbases for the stores that get generated to store the
sin and cos values into memory. This is being done by tracking symbase of the
addressof memref in the VPSubscriptInst that gets generated during VPlan
construction and reusing the same to set the symbase of the stores that are
generated in VPTransformLibraryCalls. This, however, did not work for the case
of SelfAddressOfRefs where we were avoiding generation of VPSubscriptInst. The
changes here remove this exception by generating a VPSubscriptInst without any
dimension info. Such VPSubscriptInsts are now treated as SelfAddressOf
instructions. This mainly needed changes in code that processes loop entities
which was looking for direct uses of pointer entities. Minor changes were
also needed in planner, divergence analysis, SOA analysis, and cost modeling
to account for the same.

* Address review comments.