[NFC] Move DTrans type metadata reader tests to OpaquePtr directory (#7073)
The DTrans type metadata component is needed for the new DTrans
infrastructure for opaque pointers. When it was first developed,
opaque pointer types did not exist in the compiler, and tests
for the new DTrans infrastructure were not being placed in a
separate directory. Update the tests that should be using opaque
pointer syntax, and move them to the OpaquePtr directory

[CMPLRLLVM-40051]