[OpenMP] Give a diagnostic on use of a unmarked declare target variable (#6991)
When a global used in device code is not marked 'declare target'
the current codegen leaves it as an external reference. This causes
the program to fail at runtime usually with an error such as:

error: Global constant without initializer!

We need an error for this so users can correct this problem at
compile time instead. Since we don't know which functions are
used until codegen, we can't detect this until then and therefore,
we need to give this diagnostic in codegen when it is encountered.

Give an error when the unmarked global is used, but only outside
a target construct. Marking is not need in a target construct since a
map clause, either explicit or implicit, should handle the variable.

Variables without a valid source position do not need a diagnostic
since these are only used to create implicit clauses. If a diagnostic
was needed it would happen at the real use of the variable.

Fixes: CMPLRLLVM-39384