[VPlan][VecClone][VectorVariant] Fix for cmplrllvm-22849 (#7072)
* [VPlan][VecClone][VectorVariant] Fix for cmplrllvm-22849

This patch adds support for #pragma omp declare simd uniform(c) linear(val(x):c),
(a.k.a., #pragma omp declare simd uniform(c) linear(x:c)), which is the case where
strides for linear arguments are passed via another argument. E.g., the linear integer
variable stride case, or 'ls' encoding. Previously, VecClone only supported constant
strides, the 'l' encoding. In the near future, support will be added for R/U/L
encodings, which are associated only with C++ linear reference parameters. The
DeadArgumentElimination changes are required because it's possible that there are
no explicit uses of the stride variable within the variant itself and we don't want
that argument to be eliminated because VecClone needs it to calculate the stride.

* Address review comments

* Simplify linear checks in parameter matching function

* Fix failing lit tests reported by alloy

* Address more review comments