[VPlan][HIR] Suppress dynamic peeling in HIR path if aligned clause is used with omp simd (#7196)
* [VPlan][HIR] Suppress dynamic peeling in HIR path if aligned clause is used with omp simd

VPlan HIR vectorization path needs further work to properly consume the
assumes that are generated in VPO paropt pass from aligned clauses. Until
we address this, suppress dynamic peeling if aligned clause is used. This
is done to avoid unnecessary peeling of loops with already aligned
accesses.

VPO paropt pass replaces the original pointers used in aligned clause with
null after generating alignment assumptions. Once this happens, the
AlignedClause that we obtain is empty and as a result cannot be used to
check if the aligned clause was specified. As a result, we maintain and
use a flag in WRNVecLoopNode to track that aligned clause was used.

Lit test is added to show dynamic peeling not happening in a test for
which dynamic peeling would kick in normally.

* Address review comments